+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaReconciler.java
@@ -80,6 +80,7 @@ import io.strimzi.operator.common.model.NodeUtils;
 import io.strimzi.operator.common.model.StatusDiff;
 import io.strimzi.operator.common.operator.resource.ReconcileResult;
 import io.vertx.core.Future;
+import io.vertx.core.Promise;
 import io.vertx.core.Vertx;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.common.KafkaException;
@@ -120,6 +121,7 @@ public class KafkaReconciler {
     private final PlatformFeaturesAvailability pfa;
     private final ImagePullPolicy imagePullPolicy;
     private final List<LocalObjectReference> imagePullSecrets;
+    private final List<Integer> previousNodeIds;
 
     // Objects used during the reconciliation
     /* test */ final Reconciliation reconciliation;
@@ -210,6 +212,7 @@ public class KafkaReconciler {
         this.pfa = pfa;
         this.imagePullPolicy = config.getImagePullPolicy();
         this.imagePullSecrets = config.getImagePullSecrets();
+        this.previousNodeIds = kafkaCr.getStatus() != null ? kafkaCr.getStatus().getRegisteredNodeIds() : null;
 
         this.stsOperator = supplier.stsOperations;
         this.strimziPodSetOperator = supplier.strimziPodSetOperator;
@@ -268,6 +271,7 @@ public class KafkaReconciler {
                 .compose(i -> headlessServiceEndpointsReady())
                 .compose(i -> clusterId(kafkaStatus))
                 .compose(i -> defaultKafkaQuotas())
+                .compose(i -> nodeUnregistration(kafkaStatus))
                 .compose(i -> metadataVersion(kafkaStatus))
                 .compose(i -> deletePersistentClaims())
                 .compose(i -> sharedKafkaConfigurationCleanup())
@@ -940,7 +944,59 @@ public class KafkaReconciler {
     }
 
     /**
+     * Unregisters the KRaft nodes that were removed from the Kafka cluster
+     *
+     * @param kafkaStatus   Kafka status for updating the list of currently registered node IDs
+     *
+     * @return  Future which completes when the nodes removed from the Kafka cluster are unregistered
+     */
+    protected Future<Void> nodeUnregistration(KafkaStatus kafkaStatus) {
+        List<Integer> currentNodeIds = kafka.nodes().stream().map(NodeRef::nodeId).sorted().toList();
+
+        if (kafkaMetadataStateManager.getMetadataConfigurationState().isKRaft()
+                && previousNodeIds != null
+                && !new HashSet<>(currentNodeIds).containsAll(previousNodeIds)) {
+            // We are in KRaft mode and there are some nodes that were removed => we should unregister them
+            List<Integer> nodeIdsToUnregister = new ArrayList<>(previousNodeIds);
+            nodeIdsToUnregister.removeAll(currentNodeIds);
+
+            LOGGER.infoCr(reconciliation, "Kafka nodes {} were removed from the Kafka cluster and will be unregistered", nodeIdsToUnregister);
+
+            Promise<Void> unregistrationPromise = Promise.promise();
+            KafkaNodeUnregistration.unregisterNodes(reconciliation, vertx, adminClientProvider, coTlsPemIdentity.pemTrustSet(), coTlsPemIdentity.pemAuthIdentity(), nodeIdsToUnregister)
+                    .onComplete(res -> {
+                        if (res.succeeded()) {
+                            LOGGER.infoCr(reconciliation, "Kafka nodes {} were successfully unregistered from the Kafka cluster", nodeIdsToUnregister);
+                            kafkaStatus.setRegisteredNodeIds(currentNodeIds);
+                        } else {
+                            LOGGER.warnCr(reconciliation, "Failed to unregister Kafka nodes {} from the Kafka cluster", nodeIdsToUnregister);
+
+                            // When the unregistration failed, we will keep the original registered node IDs to retry
+                            // the unregistration for them. But we will merge it with any existing node IDs to make
+                            // sure we do not lose track of them.
+                            Set<Integer> updatedNodeIds = new HashSet<>(currentNodeIds);
+                            updatedNodeIds.addAll(previousNodeIds);
+                            kafkaStatus.setRegisteredNodeIds(updatedNodeIds.stream().sorted().toList());
+                        }
+
+                        // We complete the promise with success even if the unregistration failed as we do not want to
+                        // fail the reconciliation.
+                        unregistrationPromise.complete();
+                    });
+
+            return unregistrationPromise.future();
+        } else {
+            // We are either not in KRaft mode, or at a cluster without any information about previous nodes, or without
+            // any change to the nodes => we just update the status field
+            kafkaStatus.setRegisteredNodeIds(currentNodeIds);
+            return Future.succeededFuture();
+        }
+    }
+
+    /**
      * Manages the KRaft metadata version
+     *
+     * @param kafkaStatus   Kafka status used for updating the currently used metadata version
      *
      * @return  Future which completes when the KRaft metadata version is set to the current version or updated.
      */